refactor(form-field): use lightweight token for optimized treeshaking
Currently the `MatFormField` always retains `MatHint`, `MatError`,
`MatPrefix` and `MatSuffix`. These are optional parts of a form-field so
we ideally would not retain those always.

We can achieve this by using lightweight injection tokens for querying
of these directives, instead of using the actual implementations as
query predicate token. The retention was already an issue in View Engine
too, but the issue became more noticable in Ivy where factories and
definitions are directly attached to the classes.